HomeAway - Specification review
HomeAway had specification mentioned and questions to be asked:

Question 1 - Read the specification and identify any aspect that is
unclear or risky. From this, provide a written list of questions you
would ask the Product Manager to clarify this spec before development
begins. e.g. "What should happen if the two passwords entered on the
"Edit Password" screen don't match?"

Question 2 - Imagine that the Product Manager is about to leave on a
month's holiday, and the feature will be implemented and shipped while
she is away. She is in a hurry and has to leave work to catch her flight
in 10 minutes.